pyproject.toml
[tool.poetry]
name = "test_python_manager"
version = "0.1.0"
description = ""
authors = ["shar <Sasha.sharai1998@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
djangorestframework-simplejwt = "^5.2.2"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
isort = "^5.10.1"
black = "^22.10.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
